refactoring client in progress
