+++ libethereum/Client.cpp
@@ -244,61 +244,6 @@ void Client::clearPending()
 	noteChanged(changeds);
 }
 
-unsigned Client::installWatch(h256 _h, Reaping _r)
-{
-	unsigned ret;
-	{
-		Guard l(m_filterLock);
-		ret = m_watches.size() ? m_watches.rbegin()->first + 1 : 0;
-		m_watches[ret] = ClientWatch(_h, _r);
-		cwatch << "+++" << ret << _h.abridged();
-	}
-	auto ch = logs(ret);
-	if (ch.empty())
-		ch.push_back(InitialChange);
-	{
-		Guard l(m_filterLock);
-		swap(m_watches[ret].changes, ch);
-	}
-	return ret;
-}
-
-unsigned Client::installWatch(LogFilter const& _f, Reaping _r)
-{
-	h256 h = _f.sha3();
-	{
-		Guard l(m_filterLock);
-		if (!m_filters.count(h))
-		{
-			cwatch << "FFF" << _f << h.abridged();
-			m_filters.insert(make_pair(h, _f));
-		}
-	}
-	return installWatch(h, _r);
-}
-
-bool Client::uninstallWatch(unsigned _i)
-{
-	cwatch << "XXX" << _i;
-
-	Guard l(m_filterLock);
-
-	auto it = m_watches.find(_i);
-	if (it == m_watches.end())
-		return false;
-	auto id = it->second.id;
-	m_watches.erase(it);
-
-	auto fit = m_filters.find(id);
-	if (fit != m_filters.end())
-		if (!--fit->second.refCount)
-		{
-			cwatch << "*X*" << fit->first << ":" << fit->second.filter;
-			m_filters.erase(fit);
-		}
-	return true;
-}
-
 void Client::noteChanged(h256Set const& _filters)
 {
 	Guard l(m_filterLock);
@@ -317,39 +262,6 @@ void Client::noteChanged(h256Set const& _filters)
 	// clear the filters now.
 	for (auto& i: m_filters)
 		i.second.changes.clear();
-}
-
-LocalisedLogEntries Client::peekWatch(unsigned _watchId) const
-{
-	Guard l(m_filterLock);
-
-#if ETH_DEBUG
-	cdebug << "peekWatch" << _watchId;
-#endif
-	auto& w = m_watches.at(_watchId);
-#if ETH_DEBUG
-	cdebug << "lastPoll updated to " << chrono::duration_cast<chrono::seconds>(chrono::system_clock::now().time_since_epoch()).count();
-#endif
-	w.lastPoll = chrono::system_clock::now();
-	return w.changes;
-}
-
-LocalisedLogEntries Client::checkWatch(unsigned _watchId)
-{
-	Guard l(m_filterLock);
-	LocalisedLogEntries ret;
-
-#if ETH_DEBUG && 0
-	cdebug << "checkWatch" << _watchId;
-#endif
-	auto& w = m_watches.at(_watchId);
-#if ETH_DEBUG && 0
-	cdebug << "lastPoll updated to " << chrono::duration_cast<chrono::seconds>(chrono::system_clock::now().time_since_epoch()).count();
-#endif
-	std::swap(ret, w.changes);
-	w.lastPoll = chrono::system_clock::now();
-
-	return ret;
 }
 
 void Client::appendFromNewPending(TransactionReceipt const& _receipt, h256Set& io_changed, h256 _sha3)
@@ -668,84 +580,3 @@ StateDiff Client::diff(unsigned _txi, h256 _block) const
 	return st.fromPending(_txi).diff(st.fromPending(_txi + 1));
 }
 
-LocalisedLogEntries Client::logs(unsigned _watchId) const
-{
-	LogFilter f;
-	try
-	{
-		Guard l(m_filterLock);
-		f = m_filters.at(m_watches.at(_watchId).id).filter;
-	}
-	catch (...)
-	{
-		return LocalisedLogEntries();
-	}
-	return logs(f);
-}
-
-LocalisedLogEntries Client::logs(LogFilter const& _f) const
-{
-	LocalisedLogEntries ret;
-	unsigned begin = min<unsigned>(m_bc.number() + 1, (unsigned)_f.latest());
-	unsigned end = min(m_bc.number(), min(begin, (unsigned)_f.earliest()));
-
-	// Handle pending transactions differently as they're not on the block chain.
-	if (begin > m_bc.number())
-	{
-		ReadGuard l(x_stateDB);
-		for (unsigned i = 0; i < m_postMine.pending().size(); ++i)
-		{
-			// Might have a transaction that contains a matching log.
-			TransactionReceipt const& tr = m_postMine.receipt(i);
-			auto sha3 = m_postMine.pending()[i].sha3();
-			LogEntries le = _f.matches(tr);
-			if (le.size())
-				for (unsigned j = 0; j < le.size(); ++j)
-					ret.insert(ret.begin(), LocalisedLogEntry(le[j], begin, sha3));
-		}
-		begin = m_bc.number();
-	}
-
-	set<unsigned> matchingBlocks;
-	for (auto const& i: _f.bloomPossibilities())
-		for (auto u: m_bc.withBlockBloom(i, end, begin))
-			matchingBlocks.insert(u);
-
-#if ETH_DEBUG
-	unsigned falsePos = 0;
-#endif
-	for (auto n: matchingBlocks)
-	{
-#if ETH_DEBUG
-		int total = 0;
-#endif
-		auto h = m_bc.numberHash(n);
-		auto receipts = m_bc.receipts(h).receipts;
-		for (size_t i = 0; i < receipts.size(); i++)
-		{
-			TransactionReceipt receipt = receipts[i];
-			if (_f.matches(receipt.bloom()))
-			{
-				auto info = m_bc.info(h);
-				auto h = transaction(info.hash, i).sha3();
-				LogEntries le = _f.matches(receipt);
-				if (le.size())
-				{
-#if ETH_DEBUG
-					total += le.size();
-#endif
-					for (unsigned j = 0; j < le.size(); ++j)
-						ret.insert(ret.begin(), LocalisedLogEntry(le[j], n, h));
-				}
-			}
-#if ETH_DEBUG
-			if (!total)
-				falsePos++;
-#endif
-		}
-	}
-#if ETH_DEBUG
-	cdebug << matchingBlocks.size() << "searched from" << (end - begin) << "skipped; " << falsePos << "false +ves";
-#endif
-	return ret;
-}
